This didn't improve the performance?

Before:

```sh
Benchmark 1: stak-interpret bench/fibonacci/main.out
  Time (mean ± σ):      8.995 s ±  0.102 s    [User: 8.914 s, System: 0.007 s]
  Range (min … max):    8.794 s …  9.133 s    10 runs
```

After:

```sh
Benchmark 1: stak-interpret bench/fibonacci/main.out
  Time (mean ± σ):      9.232 s ±  0.203 s    [User: 9.122 s, System: 0.008 s]
  Range (min … max):    8.964 s …  9.697 s    10 runs
```